pyproject.toml
[tool.poetry]
name = "pptx2md"
version = "1.5.0"
description = "This package converts pptx to markdown"
repository = "https://github.com/ssine/pptx2md"
authors = ["Liu Siyao <liu.siyao@qq.com>"]
license = "MIT Licence"

[tool.poetry.scripts]
pptx2md = "pptx2md.__main__:main"

[tool.poetry.dependencies]
python = ">=3.7"
python-pptx = ">=0.6.18"
rapidfuzz = ">=0.10.0"
Pillow = ">=6.0.0"
tqdm = ">=4"

[tool.poetry.dev-dependencies]
yapf = ">=0.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
column_limit = 120
